requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
numpy==1.24.3
torch==2.0.1
torch-geometric==2.3.1
pandas==1.5.3
captum==0.6.0
python-multipart==0.0.5
websockets==10.4
scipy==1.10.1
requests==2.28.2
matplotlib==3.7.1
seaborn==0.12.2
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
authlib==1.2.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2